allwinner: A64: Add AXP803 PMIC support to power off the board
Boards with the Allwinner A64 SoC are mostly paired with an AXP803 PMIC,
which allows to programmatically power down the board.
Use the existing (Mentor Graphics) I2C driver to detect and program the
PMIC connected to the R_I2C pins.
Use that connection to turn off the main voltage rails when receiving a
PSCI SYSTEM_POWER_OFF command.

Linux uses the RSB bus to drive the PMIC chip, so we try to detect this
and switch it back to using I2C before powering down the board.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>