allwinner: A64/H5: WIP! use the arisc to turn off cores
PSCI requires a core to turn itself off, which we can't do properly by
just executing an algorithm on that very core. As a consequence we just
put a core into WFI on CPU_OFF right now.
To fix this let's task the "arisc" management processor (an OpenRISC
core) with that task of asserting reset and turning off the core's power
domain. We use a handcrafted sequence of OpenRISC instructions in an
array, which is visible to the arisc as well. We put the distance to
that array in the OpenRISC reset vector and release the reset line to
let the arisc do its job. The arisc de-asserts the reset line at the end
to stop execution and prepare for the next injection.

TODO:
- let arisc code wait for the respective WFI bit in R_CPUCFG
- adjust or mask for H6
- check cache situation
- lock to avoid parallel usage (needed?)
- check for already used/running arisc?

Signed-off-by: Andre Przywara <andre.przywara@arm.com>