FIXUP! Add basic support for Raspberry Pi 4
The arch timer should really tick at 54 MHz, but I changed the value
that ends up in CNTFRQ to 2.7MHz before, to match reality.
Now by having a closer look at the MMIO addresses where we try to adjust
the arch timer prescaler, I see how it was off ;-)

Fix the stupid bug that prevented us from actually programming the
prescaler, and set the CNTFRQ to the actual value.

Change-Id: I4212a5563851f28b57b803f22369cb417967735e
Signed-off-by: Andre Przywara <andre.przywara@arm.com>